+++ locale/frFR.lua
@@ -147,6 +147,8 @@ L["SCORPID_MODE_DETAILED_DESC"] = "Ce mode détecte lorsqu'un chasseur lance la 
 L["MODE_BUTTON_DESC"] = "Affiche le bouton pour activer le mode '%s'"
 L["MODE_LABEL"] = "Texte du bouton"
 L["MODE_LABEL_DESC"] = "Texte écrit dans le bouton pour activer le mode '%s'"
+L["MODE_TRACK_FOCUS"] = "Aligne le focus avec l'affectation"
+L["MODE_TRACK_FOCUS_DESC"] = "Détecte quand votre affectation change et propose d'aligner le focus avec la nouvelle affectation"
 
     --- Names
 L["SETTING_NAMES"] = "Noms"
@@ -211,6 +213,8 @@ L["HISTORY_GROUNDING_CANCEL"] = "Le totem de %s fut annulé prématurément à c
 L["HISTORY_GROUNDING_EXPIRE"] = "Le totem de %s expire."
 L["HISTORY_GROUNDING_ABSORB"] = "Le totem de %s absorbe %s de %s."
 L["HISTORY_GROUNDING_ABSORB_NOSPELL"] = "Le totem de %s absorbe une attaque de %s."
+L["HISTORY_ASSIGN_PLAYER"] = "%s a affecté %s à se focaliser sur %s."
+L["HISTORY_ASSIGN_NOBODY"] = "%s a retiré l'affectation de %s."
 
     --- Icons
 L["DISPLAY_BLIND_ICON"] = "Afficher une icône pour les joueurs qui n'ont pas installé SilentRotate"
@@ -226,6 +230,20 @@ L["TOOLTIP_EFFECT_REMAINING"] = "Durée de l'effet : %s"
 L["TOOLTIP_COOLDOWN_REMAINING"] = "Temps de recharge : %s"
 L["TOOLTIP_DURATION_SECONDS"] = "%s sec"
 L["TOOLTIP_DURATION_MINUTES"] = "%s min"
+L["TOOLTIP_ASSIGNED_TO"] = "Affecté(e) à : %s"
+L["TOOLTIP_EFFECT_CURRENT"] = "Maintenant sur : %s"
+L["TOOLTIP_EFFECT_PAST"] = "Dernière cible : %s"
+
+    --- Context Menu
+L["CONTEXT_ASSIGN_TITLE"] = "Affecter %s à :"
+L["CONTEXT_NOBODY"] = "Personne"
+L["CONTEXT_CANCEL"] = "Annuler"
+L["CONTEXT_OTHERS"] = "Autres joueurs"
+
+    --- Dialog Box
+L["DIALOG_ASSIGNMENT_QUESTION1"] = "Votre focus diffère de votre affectation."
+L["DIALOG_ASSIGNMENT_QUESTION2"] = "Voulez-vous définir le focus sur %s?"
+L["DIALOG_ASSIGNMENT_CHANGE_FOCUS"] = "Changer le focus"
 
     --- Notifications
 L["UPDATE_AVAILABLE"] = "Une nouvelle version est disponible, veuillez mettre à jour pour profiter des dernières fonctionnalités."